+++ internal/agent/memstat/metricservice.go
@@ -1,23 +1,17 @@
 package memstat
 
 import (
-	"bytes"
 	"crypto/rand"
-	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"math/big"
-	"net/http"
 	"runtime"
 	"sync"
 	"time"
 
 	"github.com/fdanis/ygtrack/internal/constants"
-	"github.com/fdanis/ygtrack/internal/helpers"
 	"github.com/fdanis/ygtrack/internal/server/models"
 	"github.com/shirou/gopsutil/v3/mem"
-	"golang.org/x/sync/errgroup"
 )
 
 const (
@@ -25,27 +19,21 @@ const (
 	randomCount = "RandomValue"
 )
 
-type Service struct {
+type MetricService struct {
 	gaugeDictionary map[string]float64
 	countDictionary map[string]int64
-	send            func(client *http.Client, url string, header map[string]string, data *bytes.Buffer) error
 	lock            sync.RWMutex
 	hashkey         string
-	httpclient      *http.Client
-	workers         int
 }
 
-func NewService(hashkey string) *Service {
-	m := new(Service)
-	m.send = post
-	m.httpclient = &http.Client{}
+func NewMetricService(hashkey string) *MetricService {
+	m := new(MetricService)
 	m.gaugeDictionary = map[string]float64{}
 	m.countDictionary = map[string]int64{pollCount: 0}
 	m.hashkey = hashkey
-	m.workers = 10
 	return m
 }
-func (m *Service) Update() {
+func (m *MetricService) Update() {
 	fmt.Printf("update metrics %s \n", time.Now().Format("15:04:05"))
 	m.lock.Lock()
 	defer m.lock.Unlock()
@@ -89,7 +77,7 @@ func (m *Service) Update() {
 	m.countDictionary[pollCount]++
 }
 
-func (m *Service) UpdateGopsUtil() {
+func (m *MetricService) UpdateGopsUtil() {
 	fmt.Printf("update gops util metrics %s \n", time.Now().Format("15:04:05"))
 	m.lock.Lock()
 	defer m.lock.Unlock()
@@ -102,105 +90,7 @@ func (m *Service) UpdateGopsUtil() {
 	m.gaugeDictionary["CPUutilization1"] = float64(virtual.UsedPercent)
 }
 
-func (m *Service) Send(url string) {
-	metrics := m.getMetrics()
-	//generate hash
-	if m.hashkey != "" {
-		g := &errgroup.Group{}
-		for _, v := range metrics {
-			gen := helpers.HashGenerator{Object: v, Key: m.hashkey}
-			g.Go(gen.Do)
-		}
-		err := g.Wait()
-		if err != nil {
-			log.Printf("could not refresh hash  %v", err)
-			//don't send any metrics if error exists
-			return
-		}
-	}
-
-	g := errgroup.Group{}
-	recordCh := make(chan *bytes.Buffer)
-	for i := 0; i < m.workers; i++ {
-		w := &SendWorker{
-			ch: recordCh,
-			send: func(data *bytes.Buffer) error {
-				return m.send(m.httpclient, url, map[string]string{"Content-Type": "application/json"}, data)
-			},
-		}
-		g.Go(w.Do)
-	}
-
-	w := &MarshalWorker{ch: recordCh, list: metrics}
-	err := w.Do()
-	if err != nil {
-		log.Println(err)
-	}
-	close(recordCh)
-
-	err = g.Wait()
-	if err != nil {
-		log.Println(err)
-	}
-}
-
-type MarshalWorker struct {
-	list []*models.Metrics
-	ch   chan *bytes.Buffer
-}
-
-func (w *MarshalWorker) Do() error {
-	for _, item := range w.list {
-		d, err := json.Marshal(item)
-		if err != nil {
-			log.Printf("could marshal %v", err)
-			return err
-		}
-		w.ch <- bytes.NewBuffer(d)
-	}
-	return nil
-}
-
-type SendWorker struct {
-	ch   chan *bytes.Buffer
-	send func(data *bytes.Buffer) error
-}
-
-func (w *SendWorker) Do() error {
-	for data := range w.ch {
-		err := w.send(data)
-		if err != nil {
-			log.Print(err)
-		}
-	}
-	return nil
-}
-
-func (m *Service) SendBatch(url string) {
-	metrics := m.getMetrics()
-	d, err := json.Marshal(metrics)
-	if err != nil {
-		log.Printf("could not do json.marshal %v", err)
-		return
-	}
-
-	var buf bytes.Buffer
-	w := io.Writer(&buf)
-	gz := helpers.GetPool().GetWriter(w)
-	defer helpers.GetPool().PutWriter(gz)
-	_, err = gz.Write(d)
-	if err != nil {
-		log.Println(err)
-	}
-	gz.Flush()
-
-	err = post(m.httpclient, url, map[string]string{"Content-Type": "application/json", "Content-Encoding": "gzip"}, &buf)
-	if err != nil {
-		log.Println("can not send batch")
-	}
-}
-
-func (m *Service) getMetrics() []*models.Metrics {
+func (m *MetricService) GetMetrics() []*models.Metrics {
 	m.lock.RLock()
 	allmetrics := make([]*models.Metrics, 0, len(m.gaugeDictionary)+2)
 	for key, val := range m.gaugeDictionary {
@@ -213,24 +103,4 @@ func (m *Service) getMetrics() []*models.Metrics {
 	}
 	m.lock.RUnlock()
 	return allmetrics
-}
-
-func post(client *http.Client, url string, header map[string]string, data *bytes.Buffer) error {
-	r, err := http.NewRequest("POST", url, data)
-	if err != nil {
-		return err
-	}
-	for k, v := range header {
-		r.Header.Add(k, v)
-	}
-	res, err := client.Do(r)
-	if err != nil {
-		return err
-	}
-
-	defer res.Body.Close()
-	if res.StatusCode != http.StatusOK {
-		return fmt.Errorf("got wrong http status (%d)", res.StatusCode)
-	}
-	return nil
 }
